edits that made environment struct more private
Just a refactor
